gas: Disable overflow check for difference relocations.
When md_apply_fix is called, the value, pointed to by VALP, is the
current best guess for what would have been patched into the fixup, even
though the fixup is due to become a relocation.

However, for difference fixups, the computed value is not a difference,
but the actual value for the first symbol in the difference.  Performing
an overflow check, as is done upon return from md_apply_fix, does not
make any sense.

This patch disables the overflow check for fixups that are due to become
difference relocations in order to avoid the overflow check raising
invalid errors.

This patch is instead of the following upstream patch:
    https://sourceware.org/ml/binutils/2015-01/msg00006.html
The upstream patch does not apply to the avr-mainline branch due to
other changes in the branch, if those other changes are ever pushed back
upstream, then this merge issue will need to be resolved.

gas/ChangeLog:

	* config/tc-avr.c (md_apply_fix): Prevent overflow check for
	fixups that are destined to become difference relocations.